Fixed remove collab API

+++ emogo/apps/collaborator/views.py
@@ -138,6 +138,7 @@ class CollaboratorDeletionAPI(DestroyAPIView):
         """
         collaborator = Collaborator.objects.get(id=kwargs.get('pk'))
         noti = Notification.objects.filter(notification_type = 'collaborator_confirmation' , stream = collaborator.stream, from_user = self.request.user, to_user = User.objects.get(username = collaborator.phone_number) )
-        noti[0].delete()
+        if noti.__len__() > 0 :
+            noti[0].delete()
         collaborator.delete()
         return custom_render_response(status_code=status.HTTP_200_OK)
